+++ clang/docs/ReleaseNotes.md
@@ -178,6 +178,9 @@ features cannot lower the translation-unit ABI level;
 ### Attribute Changes in Clang
 
 - Clang now properly propagates attributes on class and variable templates to their redeclarations, which will result in redeclarations not interfering with diagnostics. (#GH209812)
+- Added the ``null_terminated`` attribute, which can be applied to function
+  parameters of pointer or array type (of scalar type) to indicate that the
+  function expects a null-terminated buffer.
 
 ### Improvements to Clang's diagnostics
 
@@ -531,6 +534,12 @@ features cannot lower the translation-unit ABI level;
   It can be re-enabled with the ``WarnOnLockOrderReversal`` option.
 
 #### Moved checkers
+
+#### New checkers or options
+
+- Introduced the ``alpha.core.NullTerminated`` checker to detect arrays missing
+  a null terminator passed as a parameter marked with the ``null_terminated``
+  attribute.
 
 #### Diagnostic changes